CODE-220 Refactor so that videos are in own class
